feat: Merge files `ChatListScreen` and `ChatScreen` into one
- Improved the documentation of the merged code for better understanding.
- Formatted the code using Prettier to make it more organized and readable.
