+++ src/modules/chat/screens/index.ts
@@ -1,2 +1,1 @@
-export * from './ChatListScreen';
 export * from './ChatScreen';
